Reporting finished, not in UI, but report is ready.

+++ src/main/java/info/svetlik/rb/report/pdf/MarketOperation.java
@@ -1,11 +1,12 @@
 package info.svetlik.rb.report.pdf;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Builder;
 
 @Builder(toBuilder = true)
 public record MarketOperation(OperationType operationType, LocalDate operationDate, String isin, Currency currency,
-		double totalAmount, double quantity) {
+		double totalAmount, BigDecimal quantity) {
 
 }
